Get scaling of draugh correct and convert to float

+++ Parser/Common.fs
@@ -98,7 +98,7 @@ type StaticAndVoyageRelatedData = {
     Day: int;
     Hour: int;
     Minute: int;
-    Draught: int;
+    Draught: float;
     Destination: string;
     Dte: bool;
 }
